refactor(grammarize): move random tree generation to random_tree.py
